refactor: dashboard app
